Extract syscall_write() handler
